requirements.txt
streamlit
PyPDF2==3.0.1
google.generativeai
python-dotenv
streamlit_extras
fastapi==0.104.1
python-multipart==0.0.6
uvicorn==0.24.0
google-generativeai==0.3.0
PyPDF2==3.0.1
python-dotenv==1.0.0
python-docx==1.1.2
textract==1.6.5
